app/requirements.txt
semantic-kernel
pytest-asyncio